dtype fixing for hardtanh

+++ ivy/functional/frontends/torch/nn/functional/non_linear_activation_functions.py
@@ -325,7 +325,7 @@ def hardsigmoid(input, inplace=False):
 @with_unsupported_dtypes({"1.11.0 and below": ("float16",)}, "torch")
 def hardtanh(input, min_val=-1.0, max_val=1.0, inplace=False):
     less = ivy.where(ivy.less(input, min_val), min_val, input)
-    ret = ivy.where(ivy.greater(input, max_val), max_val, less)
+    ret = ivy.where(ivy.greater(input, max_val), max_val, less).astype(input.dtype)
     if inplace:
         ivy.inplace_update(input, ret)
         return input
@@ -335,7 +335,7 @@ def hardtanh(input, min_val=-1.0, max_val=1.0, inplace=False):
 @with_unsupported_dtypes({"1.11.0 and below": ("float16",)}, "torch")
 def hardtanh_(input, min_val=-1.0, max_val=1.0):
     less = ivy.where(ivy.less(input, min_val), min_val, input)
-    ret = ivy.where(ivy.greater(input, max_val), max_val, less)
+    ret = ivy.where(ivy.greater(input, max_val), max_val, less).astype(input.dtype)
     ivy.inplace_update(input, ret)
     return input
